Omer: PCQueue implementation + test

+++ Code_Skeleton/Part2/Game.cpp
@@ -10,7 +10,7 @@ void Game::run() {
 		auto gen_start = std::chrono::system_clock::now();
 		_step(i); // Iterates a single generation 
 		auto gen_end = std::chrono::system_clock::now();
-		m_gen_hist.push_back((float)std::chrono::duration_cast<std::chrono::microseconds>(gen_end - gen_start).count());
+		m_gen_hist.push_back((double)std::chrono::duration_cast<std::chrono::microseconds>(gen_end - gen_start).count());
 		print_board(nullptr);
 	} // generation loop
 	print_board("Final Board");
